Update example to explain where tokenizer is
Dearest maintainer,

I hope things are well. I found this while looking for..well this.. I want to run the example but I could not find tokenizer. I see it is in the example file. I linked it in the comment. I would either add this link or make this example smaller and not look like one can run it as it.

Hope it helps someone else. Thanks for your time on the code and reviewing this pr.
Becker